requirements.txt
streamlit
ultralytics
google-generativeai
pillow
opencv-python
gtts
speechrecognition